requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.0
python-multipart==0.0.6
email-validator==2.1.0.post1
pydantic==2.6.1
pydantic-core==2.14.5
gunicorn==20.1.0 